requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
asdfghjkl
asdl
async-timeout
attrs
backpack
backpack-for-pytorch
bayesian-torch
certifi
charset-normalizer
contourpy
curvlinops-for-pytorch
cycler
einconv
einops
exceptiongroup
filelock
fonttools
frozenlist
fsspec
grpcio
hf-xet
huggingface-hub
idna
iniconfig
Jinja2
joblib
kiwisolver
lightning-utilities
littleutils
lmdb
Markdown
MarkupSafe
matplotlib
mpmath
multidict
networkx
numpy
ogb
opt_einsum
outdated
packaging
pandas
patsy
pillow
pluggy
propcache
protobuf
psutil
pyaml
pycalib
pyparsing
pytest
python-dateutil
pytz
PyYAML
regex
requests
safetensors
scikit-learn
scipy
simplejson
six
statsmodels
sympy
tabulate
tensorboard
tensorboard-data-server
threadpoolctl
tokenizers
tomli
torch
torch-geometric
torch-scatter
torchmetrics
torchvision
tqdm
transformers
typing_extensions
tzdata
unfoldNd
urllib3
Werkzeug
wilds
yarl
laplace-torch